[breaking] move http-adapter to jquery directory
