+++ src/http-adapter/jquery/index.js
