+++ src/index.js
@@ -1,6 +1,6 @@
-import HttpAdapter from './http-adapter'
+import JqueryHttpAdapter from './http-adapter/jquery'
 import JsonApiParser from './json-api-parser'
 import Repository from './repository'
 import Store from './store'
 
-export { HttpAdapter, JsonApiParser, Repository, Store }
+export { JqueryHttpAdapter, JsonApiParser, Repository, Store }
